% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{authorize}
\alias{authorize}
\title{Authorize R package to access endpoints}
\usage{
authorize(app_name = NULL, token = NULL, cache = FALSE, ...)
}
\arguments{
\item{token}{an output from \code{\link{oauth2.0_token}} to set as the authentication token.}

\item{cache}{Should the token be cached as an .httr-oauth file or API keys stored as global options?}

\item{...}{additional arguments to send to \code{\link{oauth2.0_token}}}
}
\value{
OAuth token saved to the environment so the package can use the users' Google data
}
\description{
This is a function to authorize the R package to access APIs interactively.
}
\examples{
\dontrun{

authorize()
}
}
