% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{automate_storage}
\alias{automate_storage}
\title{Creates GitHub Action which automatically writes to googlesheet}
\usage{
automate_storage()
}
\arguments{
\item{input}{input data to write to a googlesheet}

\item{token}{OAuth token from Google login.}

\item{gsheet}{Optionally a googlesheet to write to}

\item{overwrite}{TRUE/FALSE overwrite if there is data at the destination}

\item{append_rows}{TRUE/FALSE should the data be appended to the data?}

\item{sheet}{Index or name of the worksheet you want to write to. Forwarded to googlesheets4::write_sheet or googlesheets4::append_sheet to indicate what sheet it should be written to.}

\item{new_sheet}{default is FALSE. But if it is anything else will be used as the name for a new worksheet that will be made and written to.}

\item{...}{these parameters are sent to googlesheets4::write_sheet.}
}
\value{
The googlesheet URL where the data has been written
}
\description{
This is a function to write metricminer data to a GoogleSheet
}
\examples{
\dontrun{

}
}
