% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-analytics.R
\name{clean_metric_data}
\alias{clean_metric_data}
\title{Handle Google Analytics Lists}
\usage{
clean_metric_data(metrics)
}
\arguments{
\item{account_id}{the account id of the properties you are trying to retrieve}
}
\description{
This is a function to gets metrics and dimensions for all properties associated with an account
}
