% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_creds_path}
\alias{default_creds_path}
\title{Default Credentials path}
\usage{
default_creds_path(app_name)
}
\arguments{
\item{app_name}{What app set up are you looking for? Supported apps are 'google' 'calendly' and 'github'
Get file path to an default credentials RDS}
}
\description{
Default Credentials path
}
