% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-forms.R
\name{extract_answers}
\alias{extract_answers}
\title{Google Form handling functions -- extracting answers}
\usage{
extract_answers(form_info, metadata = NULL)
}
\arguments{
\item{form_info}{The return form_info list that is extracted in `get_google_form()`}

\item{metadata}{Optional metadata returned by `get_question_metadata()`}
}
\value{
This returns answers from a google form
}
\description{
This is a function to get extract answers from a Google Form. It is
 used by the `get_google_form()` function if dataformat = "dataframe"
}
