% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{get_config_file}
\alias{get_config_file}
\title{Get config file}
\usage{
get_config_file(overwrite = FALSE)
}
\arguments{
\item{overwrite}{Should a _config_automation.yml file in the current directory be overwritten? Default is false.}
}
\value{
Copies a config_automation.yml file to your current working directory
}
\description{
Get the _config_automation.yml file to set up a metricminer repo
}
