% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-analytics.R
\name{get_ga_property_info}
\alias{get_ga_property_info}
\title{Get all property information for a particular property id}
\usage{
get_ga_property_info(property_id, token = NULL)
}
\arguments{
\item{property_id}{the property id you want information about.}

\item{token}{credentials for access to Google using OAuth.  `authorize("google")`}
}
\value{
All the property ids and information about them for a Google Analytics account.
}
\description{
This is a function to get the Google Analytics accounts that this user has access to.
The scope it uses is the `See and download your Google Analytics data.`
If you don't check this box on the OAuth screen, this function will not work.
}
\examples{
\dontrun{

authorize("google")

property_info <- get_ga_property_info(property_id = "422558989")
}
}
