% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-analytics.R
\name{get_ga_stats}
\alias{get_ga_stats}
\title{Get stats for an associated Google Analytics property}
\usage{
get_ga_stats(
  property_id,
  start_date = "2015-08-14",
  body_params = NULL,
  end_date = NULL,
  token = NULL,
  stats_type = "metrics"
)
}
\arguments{
\item{property_id}{a GA property. Looks like '123456789' Can be obtained from running `get_ga_properties()`}

\item{start_date}{YYYY-MM-DD format of what metric you'd like to collect metrics from to start. Default is the earliest date Google Analytics were collected.}

\item{body_params}{The body parameters for the request}

\item{end_date}{YYYY-MM-DD format of what metric you'd like to collect metrics from to end. Default is today.}

\item{token}{credentials for access to Google using OAuth. `authorize("google")`}

\item{stats_type}{Do you want to retrieve metrics or dimensions?}
}
\description{
This is a function to get the Google Analytics accounts that this user has access to
}
\examples{
\dontrun{

authorize("google")
accounts <- get_ga_user()

properties_list <- get_ga_properties(account_id = accounts$id[1])

property_id <- gsub("properties/", "", properties_list$properties$name[1])
metrics <- get_ga_stats(property_id, stats_type = "metrics")
dimensions <- get_ga_stats(property_id, stats_type = "dimensions")
}
}
