% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-analytics.R
\name{get_ga_user}
\alias{get_ga_user}
\title{Get Google Analytics Accounts}
\usage{
get_ga_user(token = NULL, request_type = "GET")
}
\arguments{
\item{token}{credentials for access to Google using OAuth.  `authorize("google")`}

\item{request_type}{Is this a GET or a POST?}
}
\value{
Information about what accounts Google Analytics credentials has access to
}
\description{
This is a function to get the Google Analytics accounts that this user has access to
}
\examples{
\dontrun{

authorize("google")
get_ga_user()
}
}
