% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_github_repo}
\alias{get_github_repo}
\title{Get the repository info}
\usage{
get_github_repo(api_key, repo)
}
\arguments{
\item{api_key}{You can provide the Personal Access Token key directly or this function will attempt to grab a PAT that was stored using the `authorize("github")` function}

\item{repo}{What is the repository name? For example in the repository fhdsl/metricminer, "metricminer" is the repo name.}
}
\value{
Information regarding a github account
}
\description{
This is a function to get the GitHub user's info
}
\examples{
\dontrun{

authorize("github")
get_github_repo()
}
}
