% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googledrive.R
\name{get_google_files}
\alias{get_google_files}
\title{Get list of files from a Google Shared Drive}
\usage{
get_google_files(drive_id = NULL, type = NULL)
}
\arguments{
\item{drive_id}{ID of the drive to retrieve a list of files from.
Otherwise will grab from user's personal drive. Can be a URL like
https://drive.google.com/drive/folders/012345ABCD or just the  "012345ABCD part}

\item{type}{What type of file would you like to see? "forms" "sheets" "folder" ?}
}
\description{
This is a function to get a list of files from a Googledrive location
}
\examples{
\dontrun{

authorize("google")
get_google_files()
}
}
