% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_org_metrics}
\alias{get_org_metrics}
\title{Retrieve metrics for all repos of an organization}
\usage{
get_org_metrics(owner, token = NULL)
}
\arguments{
\item{owner}{The owner of the repository. So for `https://github.com/fhdsl/metricminer`, it would be `fhdsl`}

\item{token}{You can provide the Personal Access Token key directly or this function will attempt to grab a PAT that was stored using the `authorize("github")` function}
}
\value{
Information regarding a github account
}
\description{
This is a function to get metrics for all the repos underneath an organization
}
\examples{
\dontrun{

authorize("github")
all_repos_metrics <- get_org_metrics(owner = "fhdsl")
}

}
