% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-forms.R
\name{get_question_metadata}
\alias{get_question_metadata}
\title{Google Form handling functions}
\usage{
get_question_metadata(form_info)
}
\arguments{
\item{form_info}{The return form_info list that is extracted in `get_google_form()`}
}
\value{
This returns metadata from a google form
}
\description{
This is a function to get metadata about a Google Form. It is
 used by the `get_google_form()` function if dataformat = "dataframe"
}
