% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_timestamp_repo_metrics}
\alias{get_timestamp_repo_metrics}
\title{Get timestamp repo metrics}
\usage{
get_timestamp_repo_metrics(results, column)
}
\arguments{
\item{repo_name}{The repository name. So for `https://github.com/fhdsl/metricminer`, it would be `metricminer`}

\item{repo_metric_list}{a list containing the metrics}
}
\value{
Metrics with timestamps for a repo on GitHub
}
\description{
This is a function to get timestamp metrics for a single repos api call response
}
