% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{get_youtube_channel_stats}
\alias{get_youtube_channel_stats}
\title{Get Youtube channel stats}
\usage{
get_youtube_channel_stats(channel_id, dataformat = "dataframe")
}
\arguments{
\item{channel_id}{ID of the youtube channel to retrieve stats from.}

\item{dataformat}{How would you like the data returned to you? Default is a "dataframe" but if you'd like to see the original API list result, put "raw".
https://www.youtube.com/channel/UCBbHCj7kUogAMFyBAzzzfUw or just the  "UCBbHCj7kUogAMFyBAzzzfUw" part}
}
\description{
This is a function to retrieve statistics for a Youtube channel
}
\examples{
\dontrun{

authorize("google")
youtube_channel_stats <- get_youtube_channel_stats("UCr73I9ZEPbn-3_1CBM57QgQ")

}
}
