% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{get_youtube_stats}
\alias{get_youtube_stats}
\title{Get Youtube stats}
\usage{
get_youtube_stats(channel_id)
}
\arguments{
\item{channel_id}{ID of the youtube channel to retrieve stats from.
https://www.youtube.com/channel/UCBbHCj7kUogAMFyBAzzzfUw or just the  "UCBbHCj7kUogAMFyBAzzzfUw" part}
}
\description{
This is a function to get a list of files from a Googledrive location
}
\examples{
\dontrun{

authorize("google")
get_get_youtube_stats("UCBbHCj7kUogAMFyBAzzzfUw")
}
}
