% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{get_youtube_video_stats}
\alias{get_youtube_video_stats}
\title{Get Youtube video stats}
\usage{
get_youtube_video_stats(video_id, dataformat = "dataframe")
}
\arguments{
\item{dataformat}{How would you like the data returned to you? Default is a "dataframe" but if you'd like to see the original API list result, put "raw".}

\item{channel_id}{ID of the youtube video to retrieve stats from.
https://www.youtube.com/watch?v=YkYnni-WuaQor just the  "YkYnni-WuaQor" part that comes after the `v=` bit.}
}
\description{
This is a function to get a statistics on a Youtube video
}
\examples{
\dontrun{

authorize("google")
youtube_video_stats <- get_youtube_video_stats("YkYnni-WuaQ")

}
}
