% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{gh_repo_wrapper}
\alias{gh_repo_wrapper}
\title{Wrapper function for gh repository calls}
\usage{
gh_repo_wrapper(api_call, owner, repo, token = NULL, count = Inf)
}
\arguments{
\item{api_call}{an API call and endpoint. That has `owner` and `user`.}

\item{owner}{The repository name. So for `https://github.com/fhdsl/metricminer`, it would be `fhdsl`}

\item{repo}{The repository name. So for `https://github.com/fhdsl/metricminer`, it would be `metricminer`}

\item{token}{You can provide the Personal Access Token key directly or this function will attempt to grab a PAT that was stored using the `authorize("github")` function}

\item{count}{How many items would you like to receive? Put "all" to retrieve all records.}
}
\value{
Metrics for a repository on GitHub
}
\description{
This is a function that wraps up gh calls for us
}
