% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_example_data}
\alias{list_example_data}
\title{Get list of example datasets}
\usage{
list_example_data()
}
\description{
This is a function to retrieve a list of the example datasets included with metricminer
}
\examples{
\dontrun{

list_example_data()
}
}
