% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{me_robot}
\alias{me_robot}
\title{Check to see if the response suspects I am a robot}
\usage{
me_robot(response)
}
\arguments{
\item{response}{the response from `request_and_check()` with the HTML}
}
\value{
a boolean TRUE for it has decided I may be a robot and FALSE for it has not seemed to decide that
}
\description{
This function looks at the first header to see if my request is suspected to be from a robot
}
