% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{parse_journal_year}
\alias{parse_journal_year}
\title{Parse the response to extract journals and publication years for papers that cite your paper}
\usage{
parse_journal_year(response)
}
\arguments{
\item{response}{the response from `request_and_check()` with the HTML}
}
\value{
a named list with vector of journal names ('journal') and a vector of publication years ('pubyr')
}
\description{
This function takes the response from `request_and_check()` and parses it to extract specifically the displayed journals and publication years
}
