% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{parse_links}
\alias{parse_links}
\title{Parse the response to extract links of papers that cite your paper}
\usage{
parse_links(response)
}
\arguments{
\item{response}{the response from `request_and_check()` with the HTML}
}
\value{
a vector of links
}
\description{
This function takes the response from `request_and_check()` and parses it to extract specifically the displayed links of the papers
}
