% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{parse_titles}
\alias{parse_titles}
\title{Parse the response to extract titles of papers that cite your paper}
\usage{
parse_titles(response)
}
\arguments{
\item{response}{the response from `request_and_check()` with the HTML}
}
\value{
a vector of paper titles
}
\description{
This function takes the response from `request_and_check()` and parses it to extract specifically the displayed titles of the papers
}
