% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{request_and_check}
\alias{request_and_check}
\title{GET and check response}
\usage{
request_and_check(paper_cite_link)
}
\arguments{
\item{paper_cite_link}{the citation link provided to `get_citation_count` perhaps with a start parameter included}
}
\value{
the response or stop the process
}
\description{
This is a function that given a URL uses GET to send a request and then will check the response to make sure the status code reflects data was returned or the request was ok/successful.
}
