% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-analytics.R
\name{request_ga}
\alias{request_ga}
\title{Handler for API requests from Google Analytics}
\usage{
request_ga(token, url, query = NULL, body_params = NULL, request_type)
}
\arguments{
\item{token}{credentials for access to Google using OAuth. `authorize("google")`}

\item{url}{The endpoint URL for the request}

\item{query}{A list to be passed to query}

\item{body_params}{The body parameters for the request}

\item{request_type}{Is this a GET or a POST?}
}
\value{
An API response in the form of a list
}
\description{
This is a function that handles requests from Google Analytics
}
