% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-forms.R
\name{request_google_forms}
\alias{request_google_forms}
\title{Get Google Forms}
\usage{
request_google_forms(token, url, query = NULL, body_params = NULL)
}
\arguments{
\item{token}{credentials for access to Google using OAuth. `authorize("google")`}

\item{url}{The endpoint URL for the request}

\item{query}{A list to be passed to query}

\item{body_params}{The body parameters for the request}
}
\description{
This is a function to get the Calendly API user info
}
