% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{setup_folders}
\alias{setup_folders}
\title{Setups folder structure for metricminer}
\usage{
setup_folders(
  config_file = file.path(rprojroot::find_root(rprojroot::has_dir(".git")),
    "_config_automation.yml"),
  token = NULL
)
}
\arguments{
\item{config_file}{The file path to the _config_automation.yml file}

\item{token}{OAuth token from Google login.}
}
\value{
The googlesheet URL where the data has been written
}
\description{
This is a function to setup a folder structure for metricminer data to be saved to. It depends on and reads
Scope used for this function is the `See, edit, create, and delete only the specific Google Drive files you use with this app.`
}
\examples{
\dontrun{

authorize("google")

setup_folders(
  config_file =  "_config_automation.yml"
)
}

}
