% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{write_playlist_details}
\alias{write_playlist_details}
\title{Write playlist details from YouTube}
\usage{
write_playlist_details(playlist_id, token = NULL, outfile = NULL)
}
\arguments{
\item{playlist_id}{string, playlist ID on YouTube}

\item{token}{OAuth token from Google login.}

\item{outfile}{string, a filename to which to write results in the 'resources' folder}
}
\value{
writes a file containing the dataframe of cleaned results
}
\description{
Write playlist details from YouTube
}
\examples{
\dontrun{
# Not run
write_playlist_details(
  playlist_id = shorts_playlist_id,
  outfile = "youtube_shorts_data.tsv"
)
write_playlist_details(
  playlist_id = "PL6aYJ_0zJ4uCABkMngSYjPo_3c-nUUmio",
  outfile = "youtube_shorts_data.tsv"
)
}
}
