% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{write_to_table}
\alias{write_to_table}
\title{Writes data to a tabular file}
\usage{
write_to_table(input, file_path, overwrite, table_type)
}
\arguments{
\item{input}{input data to write to a googlesheet}

\item{file_path}{A file path where the table should be saved to}

\item{overwrite}{TRUE/FALSE overwrite if there is data at the destination}

\item{table_type}{CSV and TSV are options. CSV is default}
}
\value{
The file path where the data has been written
}
\description{
This is a function to write metricminer data to a tabular file
}
\examples{
\dontrun{

authorize("github")
repo_list <- get_user_repo_list(owner = "metricminer")

write_to_table(repo_list)
}
}
