% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-analytics.R
\name{all_ga_metrics}
\alias{all_ga_metrics}
\title{Get all metrics for all properties associated with an account}
\usage{
all_ga_metrics(account_id, format = "dataframe")
}
\arguments{
\item{account_id}{the account id of the properties you are trying to retrieve}

\item{format}{How would you like the data returned to you? Default is a "dataframe" but if you'd like to see the original API list result, put "raw".}
}
\value{
Either a list of dataframes where `metrics`, `dimensions` and `link clicks` are reported. But if `format` is set to "raw" then the original raw API results will be returned
}
\description{
This is a function to gets metrics and dimensions for all properties associated with an account
}
\examples{
\dontrun{

authorize("google")
accounts <- get_ga_user()

stats_list <- all_ga_metrics(account_id = accounts$id[5])
}
}
