% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-analytics.R
\name{clean_ga_metrics}
\alias{clean_ga_metrics}
\title{Handle Google Analytics Lists}
\usage{
clean_ga_metrics(metrics = NULL)
}
\arguments{
\item{metrics}{a metrics object from `get_ga_stats()` function}
}
\value{
a data frame of cleaned metrics from Google Analytics
}
\description{
These functions are to clean metric and dimension data from Google Analytics `get_ga_stats()` function
}
