% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-analytics.R
\name{clean_metric_data}
\alias{clean_metric_data}
\title{Handle Google Analytics Lists}
\usage{
clean_metric_data(metrics = NULL)
}
\arguments{
\item{metrics}{a metrics object from `get_ga_stats()` function}
}
\description{
These functions are to clean metric and dimension data from Google Analytics `get_ga_stats()` function
}
