% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{example_data_folder}
\alias{example_data_folder}
\title{Default creds path
Get file path to an default credentials RDS}
\usage{
example_data_folder()
}
\description{
Default creds path
Get file path to an default credentials RDS
}
