% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendly.R
\name{get_calendly_user}
\alias{get_calendly_user}
\title{Get Calendly API user}
\usage{
get_calendly_user(token = NULL)
}
\arguments{
\item{token}{You can provide the API key directly using this argument or this function will attempt to grab an API key that was stored using the `authorize("calendly")` function}
}
\value{
Calendly API user info as a list
}
\description{
This is a function to get the Calendly API user info
}
\examples{
\dontrun{

authorize("calendly")
get_calendly_user()
}
}
