% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{get_citation_count}
\alias{get_citation_count}
\title{Get a list of papers that cite your paper}
\usage{
get_citation_count(paper_cite_link)
}
\arguments{
\item{paper_cite_link}{This is not a google citation page. 1. Go to: https://scholar.google.com/scholar 2. Search for the paper we are looking for the citation count. 3. Then click the Cited by ___ button below the title of the paper 4. Copy and paste this url and put it in this get_citation_count() function}
}
\value{
A list of the example datasets available in this package
}
\description{
This is a function to retrieve a list of papers that cite your papers
}
\examples{
\dontrun{

paper_cite_link <- "https://scholar.google.com/scholar?cites=6140457238337460780"

papers_cited_df <- get_citation_count(paper_cite_link)
}
}
