% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_example_data}
\alias{get_example_data}
\title{Get retrieve an example dataset}
\usage{
get_example_data(dataset_name, envir = 1)
}
\arguments{
\item{dataset_name}{the name of the example dataset to be retrieved from the metricminer package.}

\item{envir}{By default the example data is saved in the global environment but this parameter allows you to change that if desired.}
}
\value{
an object in the environment of the same example dataset name that was requested.
}
\description{
This is a function to retrieve a list of the example datasets included with metricminer
}
\examples{
\dontrun{

# You can see the list of example datasets by running:
list_example_data()

# Then use the datasetes of your interest by calling it with this function
get_example_data("gform_info")

# Then if you check your global environment you will see "gform_info" included
ls()
}
}
