% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_github_metrics}
\alias{get_github_metrics}
\title{Get the repository metrics}
\usage{
get_github_metrics(repo, token = NULL)
}
\arguments{
\item{repo}{The repository name. So for `https://github.com/fhdsl/metricminer`, it would be `fhdsl/metricminer`}

\item{token}{You can provide the Personal Access Token key directly or this function will attempt to grab a PAT that was stored using the `authorize("github")` function}
}
\value{
Information regarding a github account
}
\description{
This is a function to get the information about a repository
}
\examples{
\dontrun{

authorize("github")
get_github_metrics(repo = "fhdsl/metricminer")
}
}
