% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{get_github_repo}
\alias{get_github_repo}
\title{Get the repository info}
\usage{
get_github_repo(token, owner, repo)
}
\arguments{
\item{token}{You can provide the Personal Access Token key directly or this function will attempt to grab a PAT that was stored using the `authorize("github")` function}

\item{owner}{The owner}

\item{repo}{The repository}
}
\value{
Information regarding a github account
}
\description{
This is a function to get the GitHub user's info
}
\examples{
\dontrun{

authorize("github")
get_github_repo()
}
}
