% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{get_youtube_channel_stats}
\alias{get_youtube_channel_stats}
\title{Get Youtube channel stats}
\usage{
get_youtube_channel_stats(channel_id, token = NULL, dataformat = "dataframe")
}
\arguments{
\item{channel_id}{ID of the Youtube channel to retrieve stats from.}

\item{token}{OAuth token from Google login.}

\item{dataformat}{How would you like the data returned to you? Default is a "dataframe" but if you'd like to see the original API list result, put "raw".}
}
\value{
A data frame of the channel stats from a Youtube channel.
}
\description{
This is a function to retrieve statistics for a Youtube channel
}
\examples{
\dontrun{

authorize("google")
youtube_channel_stats <- get_youtube_channel_stats("UCr73I9ZEPbn-3_1CBM57QgQ")
}
}
