% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youtube.R
\name{get_youtube_video_stats}
\alias{get_youtube_video_stats}
\title{Get Youtube video stats}
\usage{
get_youtube_video_stats(video_id, token = NULL, dataformat = "dataframe")
}
\arguments{
\item{video_id}{ID of the Youtube video to retrieve stats from.}

\item{token}{OAuth token from Google login.
https://www.youtube.com/watch?v=YkYnni-WuaQor just the  "YkYnni-WuaQor" part that comes after the `v=` bit.}

\item{dataformat}{How would you like the data returned to you? Default is a "dataframe" but if you'd like to see the original API list result, put "raw".}
}
\value{
A data frame of the Youtube video stats.
}
\description{
This is a function to get a statistics on a Youtube video
}
\examples{
\dontrun{

authorize("google")
youtube_video_stats <- get_youtube_video_stats("YkYnni-WuaQ")
}
}
