% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_example_data}
\alias{list_example_data}
\title{Get list of example datasets}
\usage{
list_example_data()
}
\value{
A list of the example datasets available in this package
}
\description{
This is a function to retrieve a list of the example datasets included with metricminer
}
\examples{
\dontrun{

list_example_data()

# Now you could use any of these example datasets that are printed out

get_example_data("calendly_events")

}
}
