% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-forms.R
\name{request_google_forms}
\alias{request_google_forms}
\title{Get Google Forms}
\usage{
request_google_forms(
  token,
  url,
  body_params = NULL,
  query_params = NULL,
  return_request = TRUE
)
}
\arguments{
\item{token}{credentials for access to Google using OAuth. `authorize("google")`}

\item{url}{The endpoint URL for the request}

\item{body_params}{The body parameters for the request}

\item{query_params}{The body parameters for the request}

\item{return_request}{Should a list of the request be returned as well?}
}
\value{
This function returns a list from a API response JSON file
}
\description{
This is a function to get the Calendly API user info
}
