% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{write_to_gsheet}
\alias{write_to_gsheet}
\title{Writes data to a Googlesheet}
\usage{
write_to_gsheet()
}
\arguments{
\item{input}{input data to write to a googlesheet}

\item{gsheet}{Optionally a googlesheet to write to}

\item{overwrite}{TRUE/FALSE overwrite if there is data at the destination}
}
\value{
The googlesheet URL where the data has been written
}
\description{
This is a function to write metricminer data to a GoogleSheet
}
\examples{
\dontrun{

authorize("github")
repo_list <- get_user_repo_list(owner = "metricminer")

write_to_gsheet(repo_list)
}

}
